Create and apply movies pinia store
