+++ src/components/comingSoon/ComingSoonMovies.vue
@@ -1,5 +1,7 @@
 <script>
 import { defineComponent } from "vue";
+import { mapState } from "pinia";
+import { useMoviesStore } from "@/stores/movies";
 import BaseMovieCard from "@/components/global/BaseMovieCard.vue";
 import LoadingData from "@/components/global/LoadingData.vue";
 import ErrorMessage from "@/components/global/ErrorMessage.vue";
@@ -7,15 +9,15 @@ import ErrorMessage from "@/components/global/ErrorMessage.vue";
 export default defineComponent({
   name: "ComingSoonMovies",
   components: { BaseMovieCard, LoadingData, ErrorMessage },
-  props: {
-    moviesIsLoading: {
-      type: Boolean,
-    },
-    moviesErrorMessage: {
-      type: String,
-    },
-    moviesComingSoon: {
-      type: Array,
+  computed: {
+    ...mapState(useMoviesStore, [
+      "movies",
+      "moviesIsLoading",
+      "moviesError",
+      "getErrorMessage",
+    ]),
+    moviesComingSoon() {
+      return this.movies.slice(0, 3);
     },
   },
 });
@@ -24,9 +26,7 @@ export default defineComponent({
 <template>
   <div class="soon__movies">
     <LoadingData v-if="moviesIsLoading" />
-    <ErrorMessage v-else-if="moviesErrorMessage">{{
-      error.message
-    }}</ErrorMessage>
+    <ErrorMessage v-else-if="moviesError">{{ getErrorMessage }}</ErrorMessage>
     <div v-else class="soon__movies-cards">
       <BaseMovieCard
         v-for="movie in moviesComingSoon"
