+++ src/pages/HomePage.vue
@@ -1,6 +1,7 @@
 <script>
 import { defineComponent } from "vue";
-import { getAllMovies } from "@/api/services/Movies";
+import { mapActions } from "pinia";
+import { useMoviesStore } from "@/stores/movies";
 import WelcomeSection from "@/components/welcome/WelcomeSection.vue";
 import ComingSoonSection from "@/components/comingSoon/ComingSoonSection.vue";
 import ScreeningsSection from "@/components/screenings/ScreeningsSection.vue";
@@ -14,30 +15,8 @@ export default defineComponent({
     ScreeningsSection,
     ContactUsSection,
   },
-  data() {
-    return {
-      moviesIsLoading: false,
-      moviesErrorMessage: null,
-      movies: [],
-    };
-  },
-  computed: {
-    moviesComingSoon() {
-      return this.movies.slice(0, 3);
-    },
-  },
   methods: {
-    async getMovies() {
-      this.moviesIsLoading = true;
-      try {
-        const respData = await getAllMovies();
-        this.movies = respData.data;
-      } catch (error) {
-        this.moviesErrorMessage = error.message;
-      } finally {
-        this.moviesIsLoading = false;
-      }
-    },
+    ...mapActions(useMoviesStore, ["getMovies"]),
   },
   mounted() {
     this.getMovies();
@@ -48,16 +27,8 @@ export default defineComponent({
 <template>
   <div>
     <WelcomeSection />
-    <ComingSoonSection
-      :moviesIsLoading="moviesIsLoading"
-      :moviesErrorMessage="moviesErrorMessage"
-      :moviesComingSoon="moviesComingSoon"
-    />
-    <ScreeningsSection
-      :moviesIsLoading="moviesIsLoading"
-      :moviesErrorMessage="moviesErrorMessage"
-      :movies="movies"
-    />
+    <ComingSoonSection />
+    <ScreeningsSection />
     <ContactUsSection />
   </div>
 </template>
